boot-hack: no real reason to use short for magic_off (

+++ src/boot-hack.c
@@ -345,9 +345,10 @@ static void nop( void ) { }
 void menu_init( void ) __attribute__((weak,alias("nop")));
 void debug_init( void ) __attribute__((weak,alias("nop")));
 
-static unsigned short int magic_off = 0; // Set to 1 to disable ML
-static unsigned short int magic_off_request = 0;
-unsigned short int magic_is_off() 
+static int magic_off = 0; // Set to 1 to disable ML
+static int magic_off_request = 0;
+
+int magic_is_off() 
 {
     return magic_off; 
 }
