+++ src/boot-hack.h
@@ -1,7 +1,7 @@
 #ifndef __BOOT_HACK_H_
 #define __BOOT_HACK_H_
 
-unsigned short int magic_is_off();
+int magic_is_off();
 void _disable_ml_startup();
 
 /* Blocks execution until config is read */
